requirements.txt
fastapi
uvicorn[standard]
pytest
httpx
pytest-cov
google-cloud-bigquery
google-adk
python-dotenv
